k-animator-project/layerworker/requirements.txt
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
azure-storage-blob==12.19.1
requests==2.31.0
numpy==1.24.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0
Wand==0.6.11
psd-tools==1.9.32
